ci: add workflow for checking module integration
For now, there is no testing for the tarantool project verifying its
integration with different modules and connectors from the ecosystem.
This is a quite huge gap in our CI system that is going to be covered
by these changes.

This patch introduces a couple of new workflow files that are named
'check_module_integration.yml' and 'build_tarantool.yml'. The main
workflow is 'check_module_integration.yml' that will run automatically
per push to master and release branches. Also, this workflow can be run
manually against a development branch. The 'build_tarantool.yml'
workflow is triggered by the main workflow and builds needed tarantool
packages (at this moment for Ubuntu Focal Fossa only), then stores them
as a build artifact. After that, the main workflow triggers the special
workflow 'artifact_testing.yml' from a module/connector project that
tarantool is going to verify integration with. The special module
workflow should download the tarantool build artifact, run tests
against it, and report the result to the main workflow. The basic
scheme describing the verification process is represented below:

check_module_integration.yml
             |
             |
             v
      build_tarantool --> build_tarantool.yml --> artifact
             |                                       |
             |                                       |
             v                                       |
     run_module_tests --> artifact_testing.yml <-----+
             |                    |
             |                    |
             |  report result     v
             + <------------- run_tests
             |
             |
             |
             v
            exit

For now, we start only with the vshard module. In the future, we are
going to extend the module list incrementally.